refactor: split admin endpoints
